#8.15 Conclusions

+++ src/controllers/videoController.js
@@ -1,11 +1,9 @@
 import Video from '../models/Video';
 import User from '../models/User';
-
 export const home = async (req, res) => {
   const videos = await Video.find({}).sort({ createdAt: 'desc' });
   return res.render('home', { pageTitle: 'Home', videos });
 };
-
 export const watch = async (req, res) => {
   const { id } = req.params;
   const video = await Video.findById(id).populate('owner');
@@ -17,19 +15,31 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found.' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
   }
   return res.render('edit', { pageTitle: `Edit: ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found.' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -38,11 +48,9 @@ export const postEdit = async (req, res) => {
   });
   return res.redirect(`/videos/${id}`);
 };
-
 export const getUpload = (req, res) => {
   return res.render('upload', { pageTitle: 'Upload Video' });
 };
-
 export const postUpload = async (req, res) => {
   const {
     user: { _id },
@@ -50,13 +58,16 @@ export const postUpload = async (req, res) => {
   const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
       fileUrl,
       owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect('/');
   } catch (error) {
     console.log(error);
@@ -69,10 +80,19 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render('404', { pageTitle: 'Video not found.' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect('/');
 };
-
 export const search = async (req, res) => {
   const { keyword } = req.query;
   let videos = [];
